requirements.txt
asgiref==3.7.2
blessed==1.20.0
bpython==0.24
certifi==2023.7.22
charset-normalizer==3.3.1
curtsies==0.4.2
cwcwidth==0.1.9
Django==4.2.6
djangorestframework==3.14.0
greenlet==3.0.1
idna==3.4
Pillow==10.1.0
Pygments==2.16.1
pytz==2023.3.post1
pyxdg==0.28
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
wcwidth==0.2.8
